+++ flutter_navigation_generator/lib/src/code_builder/import_builder.dart
@@ -2,6 +2,7 @@ import 'package:code_builder/code_builder.dart';
 import 'package:flutter_navigation_generator/src/models/importable_type.dart';
 import 'package:flutter_navigation_generator/src/models/route_config.dart';
 import 'package:flutter_navigation_generator/src/utils/utils.dart';
+import 'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart';
 
 class ImportBuilder {
   final Set<RouteConfig> routes;
@@ -9,6 +10,7 @@ class ImportBuilder {
   final ImportableType? pageType;
   final List<ImportableType> defaultGuards;
   final bool ignoreKeysByDefault;
+  final IncludeQueryParametersType? includeQueryParametersNavigatorConfig;
 
   ImportBuilder({
     required this.routes,
@@ -16,6 +18,7 @@ class ImportBuilder {
     this.pageType,
     this.ignoreKeysByDefault = true,
     this.defaultGuards = const [],
+    this.includeQueryParametersNavigatorConfig,
   });
 
   Iterable<Directive> generate() {
@@ -24,6 +27,9 @@ class ImportBuilder {
       'dart:convert',
       if (routes.any((route) => route.guards?.isNotEmpty == true) || defaultGuards.isNotEmpty)
         'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart',
+      if (routes.any((route) => route.includeQueryParameters == IncludeQueryParametersType.onlyOnWeb) ||
+          (includeQueryParametersNavigatorConfig == IncludeQueryParametersType.onlyOnWeb && routes.any((route) => route.includeQueryParameters == null)))
+        'package:flutter/foundation.dart',
     };
     imports.add(typeRefer(pageType, targetFile: targetFile).url);
     imports.addAll(defaultGuards.map((e) => typeRefer(e, targetFile: targetFile).url));