+++ flutter_navigation_generator/lib/src/code_builder/library_builder.dart
@@ -4,6 +4,7 @@ import 'package:flutter_navigation_generator/src/code_builder/navigator_builder.
 import 'package:flutter_navigation_generator/src/code_builder/route_names_builder.dart';
 import 'package:flutter_navigation_generator/src/models/importable_type.dart';
 import 'package:flutter_navigation_generator/src/models/route_config.dart';
+import 'package:flutter_navigation_generator_annotations/flutter_navigation_generator_annotations.dart';
 
 class LibraryGenerator {
   final Set<RouteConfig> routes;
@@ -14,10 +15,12 @@ class LibraryGenerator {
   final List<String> removeSuffixes;
   final List<ImportableType> defaultGuards;
   final bool ignoreKeysByDefault;
+  final IncludeQueryParametersType includeQueryParametersNavigatorConfig;
 
   LibraryGenerator({
     required this.routes,
     required this.className,
+    required this.includeQueryParametersNavigatorConfig,
     this.targetFile,
     this.pageType,
     this.unknownRoute,
@@ -37,6 +40,7 @@ class LibraryGenerator {
             targetFile: targetFile,
             defaultGuards: defaultGuards,
             ignoreKeysByDefault: ignoreKeysByDefault,
+            includeQueryParametersNavigatorConfig: includeQueryParametersNavigatorConfig,
           ).generate(),
         )
         ..body.addAll(
@@ -49,6 +53,7 @@ class LibraryGenerator {
               unknownRoute: unknownRoute,
               defaultGuards: defaultGuards,
               ignoreKeysByDefault: ignoreKeysByDefault,
+              includeQueryParametersNavigatorConfig: includeQueryParametersNavigatorConfig,
             ).generate(),
             RouteNamesBuilder(
               routes: routes,