+++ flutter_navigation_generator/lib/src/extensions/navigation_type_extension.dart
@@ -3,6 +3,7 @@ import 'package:flutter_navigation_generator_annotations/flutter_navigation_gene
 extension NavigationTypeExtension on NavigationType? {
   bool get isAsync => switch (this) {
         NavigationType.push ||
+        NavigationType.pushNotNamed ||
         NavigationType.restorablePush ||
         NavigationType.bottomSheet ||
         NavigationType.dialog ||